+++ slave/index.py
@@ -15,14 +15,23 @@ BLUE = 37
 
 master_ip = '192.168.1.100'
 
+
+
 def GPIO_init():
 	GPIO.setmode(GPIO.BOARD)
 	GPIO.setup(RED, GPIO.OUT)
-	GPIO.setup(GREEN, GPIO.OUT)
 	GPIO.setup(BLUE, GPIO.OUT)
+	GPIO.output(RED,GPIO.HIGH)
+	GPIO.output(BLUE,GPIO.HIGH)
+	GPIO.setup(GREEN, GPIO.OUT)
+	p = GPIO.PWM(GREEN,100)
+	p.start(100)
+
 
 def gpio_init(func):
 	def func():
+		GPIO_init()
+		GPIO.cleanup()
 		GPIO_init()
 		func()
 	
@@ -31,13 +40,13 @@ def gpio_init(func):
 #GPIO.cleanup()
 @gpio_init
 def GPIO_cleanup():
-	GPIO.output(RED,GPIO.LOW)
-	GPIO.output(GREEN,GPIO.LOW)
-	GPIO.output(BLUE,GPIO.LOW)
+	GPIO.output(RED,GPIO.HIGH)
+	GPIO.output(GREEN,GPIO.HIGH)
+	GPIO.output(BLUE,GPIO.HIGH)
 
 
 def on():
-	cmd = 'sudo python on.py &'
+	cmd = 'python on.py &'
 	os.system(cmd)
 
 def run():
@@ -45,16 +54,21 @@ def run():
 	os.system(cmd)
 
 def off():
-	cmd = 'sudo python off.py &'
+	cmd = 'python off.py &'
 	os.system(cmd)
 
 #the function of killing the process
 def kill_process(process):
 	ps_cmd = 'ps -ef | grep python | grep' + ' ' + process
-	for i in os.popen(kill_cmd).readlines():
-		pid = i.split()[1]
-		kill_cmd = 'kill -9' + ' ' + pid
-		os.system(kill_cmd)
+	lists = os.popen(ps_cmd).readlines()
+	for i in lists:
+		print i
+		tmp = i.split()
+		if 'grep' not in tmp:
+			pid = tmp[1]
+			kill_cmd = 'kill -9' + ' ' + pid
+			print kill_cmd
+			os.system(kill_cmd)
 
 def kill_other(process):
 	if len(os.popen('ps -ef | grep python | grep {process}'.format(process = process)).readlines()) < 2:
@@ -87,7 +101,8 @@ def main():
 					
 						else:
 							kill_process('run.py')
-							GPIO_cleanup()
+							GPIO.cleanup()
+							GPIO_init()
 						
 						#kill the process of 'off' status
 						if len(os.popen('ps -ef | grep python | grep off.py').readlines()) < 2:
@@ -95,8 +110,8 @@ def main():
 						
 						else:
 							kill_process('off.py')
-							GPIO_cleanup()
-					
+							GPIO.cleanup()
+							GPIO_init()
 						on()
 
 					else:
@@ -116,8 +131,10 @@ def main():
 							pass
 
 						else:
+							print 'in'
 							kill_process('on.py')
-							GPIO_cleanup()
+							GPIO.cleanup()
+							GPIO_init()
 
 						#kill the process of 'off' status
 						if len(os.popen('ps -ef | grep python | grep off.py').readlines()) < 2:
@@ -125,7 +142,8 @@ def main():
 						
 						else:
 							kill_process('off.py')
-							GPIO_cleanup()
+							GPIO.cleanup()
+							GPIO_init()
 
 						run()
 
@@ -142,20 +160,22 @@ def main():
 				if len(os.popen('ps -ef | grep python | grep off.py').readlines()) < 2:
 					
 					#kill the process of 'on' status
-					if len(os.popen('ps -ef | grep python | grep on.py').readlines()) < 2:
+					if len(os.popen('ps -ef | grep pytho | grep on.py').readlines()) < 2:
 						pass
 
 					else:
 						kill_process('on.py')
-						GPIO_cleanup()
+						GPIO.cleanup()
+						GPIO_init()
 
 					#kill the process of 'run' status
-					if len(os.popen('ps -ef | grep python | grep run.py').readlines()) < 2:
+					if len(os.popen('ps -ef | grep pytho | grep run.py').readlines()) < 2:
 						pass
 
 					else:
 						kill_process('run.py')
-						GPIO_cleanup()
+						GPIO.cleanup()
+						GPIO_init()
 
 					off()
 
@@ -164,7 +184,8 @@ def main():
 
 	except Exception, e:
 		with open('monitor.log', 'a') as log:
-			log.write(e)
+#			log.write(e)
+			pass
 
 if __name__ == '__main__':
 	main()
